Refactor constant package + using anyInt()
