+++ demo-db-to-kafka/src/main/java/com/ittovative/demodbtokafka/config/KafkaProducerConfig.java
@@ -11,14 +11,14 @@ import org.springframework.kafka.core.ProducerFactory;
 import java.util.HashMap;
 import java.util.Map;
 
-import static com.ittovative.demodbtokafka.constant.Kafka.BOOTSTRAP_SERVERS;
-import static com.ittovative.demodbtokafka.constant.Producer.ACKS;
-import static com.ittovative.demodbtokafka.constant.Producer.BATCH_SIZE;
-import static com.ittovative.demodbtokafka.constant.Producer.BUFFER_MEMORY;
-import static com.ittovative.demodbtokafka.constant.Producer.KEY_SERIALIZER;
-import static com.ittovative.demodbtokafka.constant.Producer.LINGER_MS;
-import static com.ittovative.demodbtokafka.constant.Producer.RETRIES;
-import static com.ittovative.demodbtokafka.constant.Producer.VALUE_SERIALIZER;
+import static com.ittovative.demodbtokafka.constant.KafkaConstant.BOOTSTRAP_SERVERS;
+import static com.ittovative.demodbtokafka.constant.ProducerConstant.ACKS;
+import static com.ittovative.demodbtokafka.constant.ProducerConstant.BATCH_SIZE;
+import static com.ittovative.demodbtokafka.constant.ProducerConstant.BUFFER_MEMORY;
+import static com.ittovative.demodbtokafka.constant.ProducerConstant.KEY_SERIALIZER;
+import static com.ittovative.demodbtokafka.constant.ProducerConstant.LINGER_MS;
+import static com.ittovative.demodbtokafka.constant.ProducerConstant.RETRIES;
+import static com.ittovative.demodbtokafka.constant.ProducerConstant.VALUE_SERIALIZER;
 
 /**
  * Configuration for the Kafka producer.
