+++ demo-db-to-kafka/src/main/java/com/ittovative/demodbtokafka/constant/SchedulerConstant.java
@@ -1,9 +1,9 @@
 package com.ittovative.demodbtokafka.constant;
 
-public final class Scheduler {
+public final class SchedulerConstant {
     public static final int FIXED_DELAY = 200;
 
-    private Scheduler() {
+    private SchedulerConstant() {
 
     }
 }
